RavenDB-24960 Fix validation and cleanup
