+++ src/Raven.Studio/typescript/viewmodels/database/tasks/editEmbeddingsGenerationTask.ts
@@ -30,6 +30,10 @@ import generalUtils = require("common/generalUtils");
 
 const minimumCommunityDeleteFrequencyInSec = TimeInSeconds.TimeInSeconds.Day * 36;
 
+const {
+    aiConnectionStringUtils: { getConnectorType, mapAiConnectionStringToSettingsDto },
+} = aiConnectionStringUtils;
+
 class editEmbeddingsGenerationTask extends shardViewModelBase {
     
     view = require("views/database/tasks/editEmbeddingsGenerationTask.html");
@@ -365,7 +369,7 @@ class editEmbeddingsGenerationTask extends shardViewModelBase {
             return;
         }
 
-        new testAiConnectionStringCommand(this.db, aiConnectionStringUtils.getConnectorType(connectionString), "TextEmbeddings", aiConnectionStringUtils.mapAiConnectionStringToSettingsDto(connectionString))
+        new testAiConnectionStringCommand(this.db, getConnectorType(connectionString), "TextEmbeddings", mapAiConnectionStringToSettingsDto(connectionString))
             .execute()
             .done((testResult) => this.testConnectionResult(testResult))
             .always(() => this.spinners.testConnection(false));
